third_party/WebKit/LayoutTests/inspector-enabled/sources/debugger/linkifier.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>// It is important that script starts on line 5 (zero-based 4)

function nonFormattedFunction() { var  i = 2 + 2; var a = 4; return a + i; }

function dummyScript()
{
    console.log("dummy string");
}

function onload()
{
    if (window.testRunner) {
        testRunner.waitUntilDone();
        testRunner.showWebInspector();
    }
    runTest();
}

function test()
{
    var scriptFormatter;
    var linkifier;
    var link;
    var script;
    var uiSourceCode;

    InspectorTest.scriptFormatter().then(startDebuggerTest);

    function startDebuggerTest(sf)
    {
        scriptFormatter = sf;
        InspectorTest.startDebuggerTest(waitForScripts);
    }

    function waitForScripts()
    {
        InspectorTest.showScriptSource("linkifier.html", debuggerTest);
    }

    function debuggerTest()
    {
        var scripts = InspectorTest.debuggerModel.scripts;
        for (var scriptId in scripts) {
            var scriptCandidate = scripts[scriptId];
            if (scriptCandidate.sourceURL === InspectorTest.mainTarget.inspectedURL() && scriptCandidate.lineOffset === 4) {
                script = scriptCandidate;
                break;
            }
        }

        uiSourceCode = Workspace.workspace.uiSourceCodeForURL(InspectorTest.mainTarget.inspectedURL());
        var linkifyMe = "at triggerError (http://localhost/show/:22:11)";
        var fragment = Components.linkifyStringAsFragment(linkifyMe);
        var anchor = fragment.querySelector('a');
        InspectorTest.addResult("The string \"" + linkifyMe + " \" linkifies to url: " + anchor.href);
        var info = Components.Linkifier._linkInfo(anchor);
        InspectorTest.addResult("The lineNumber is " + (info && info.lineNumber));
        InspectorTest.addResult("The columnNumber is " + (info && info.columnNumber));

        linkifier = new Components.Linkifier();
        var count1 = liveLocationsCount();
        link = linkifier.linkifyScriptLocation(SDK.targetManager.mainTarget(), null, InspectorTest.mainTarget.inspectedURL(), 8, 0, "dummy-class");
        var count2 = liveLocationsCount();

        InspectorTest.addResult("listeners added on raw source code: " + (count2 - count1));
        InspectorTest.addResult("original location: " + link.textContent);
        InspectorTest.addSniffer(Sources.ScriptFormatterEditorAction.prototype, "_updateButton", uiSourceCodeScriptFormatted);
        scriptFormatter._toggleFormatScriptSource();
    }

    function uiSourceCodeScriptFormatted()
    {
        InspectorTest.addResult("pretty printed location: " + link.textContent);
        scriptFormatter._discardFormattedUISourceCodeScript(UI.panels.sources.visibleView.uiSourceCode());
        InspectorTest.addResult("reverted location: " + link.textContent);

        var count1 = liveLocationsCount();
        linkifier.reset();
        var count2 = liveLocationsCount();

        InspectorTest.addResult("listeners removed from raw source code: " + (count1 - count2));

        InspectorTest.completeDebuggerTest();
    }

    function liveLocationsCount()
    {
        return Bindings.debuggerWorkspaceBinding._ensureInfoForScript(script)._locations.size;
    }
}

</script>
</head>

<body onload="onload()">
<p>
Tests that Linkifier works correctly.
<p>

</body>
</html>
