third_party/WebKit/LayoutTests/inspector-enabled/sources/debugger/linkifier-expected.txt

Tests that Linkifier works correctly.

The string "at triggerError (http://localhost/show/:22:11) " linkifies to url: http://localhost/show/
The lineNumber is 21
The columnNumber is 10
listeners added on raw source code: 1
original location: linkifier.html:9
pretty printed location: linkifier.html:formatted:14
reverted location: linkifier.html:9
listeners removed from raw source code: 1
